requirements.txt
aiofiles==24.1.0
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
Authlib==1.6.5
blinker==1.9.0
Brotli==1.1.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.7
cycler==0.12.1
cyclopts==3.24.0
databricks-sdk==0.67.0
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.17.0
docutils==0.22.2
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.119.0
fastmcp==2.12.4
ffmpy==0.6.3
filelock==3.20.0
Flask==3.1.2
fonttools==4.60.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
grad-cam==1.5.5
gradio==5.49.1
gradio_client==1.13.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy-object-proxy==1.12.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mcp==1.17.0
mdurl==0.1.2
mlflow==3.4.0
mlflow-skinny==3.4.0
mlflow-tracing==3.4.0
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opentelemetry-api==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
packaging==25.0
pandas==2.3.3
parse==1.20.2
pathable==0.4.4
pillow==11.3.0
pluggy==1.6.0
protobuf==6.32.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
pyperclip==1.11.0
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rfc3339-validator==0.1.4
rich==14.2.0
rich-rst==1.3.1
rpds-py==0.27.1
rsa==4.9.1
ruff==0.14.0
safehttpx==0.1.6
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.3
sse-starlette==3.0.2
starlette==0.48.0
sympy==1.14.0
threadpoolctl==3.6.0
tomli==2.3.0
tomlkit==0.13.3
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
ttach==0.0.3
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
waitress==3.0.2
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.1
zipp==3.23.0
